pyproject.toml
[project]
name = "churn-intelligence-platform"
version = "1.0.0"
description = "Agentic MLOps system for customer churn prediction with MCP, MLflow, DVC, and SHAP"
requires-python = ">=3.12"
dependencies = [
    "dvc>=3.56.0",
    "fastmcp>=2.14.4",
    "joblib>=1.4.2",
    "matplotlib>=3.9.0",
    "mlflow>=2.16.0",
    "numba>=0.60.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.5.0",
    "shap>=0.46.0",
    "xgboost>=2.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
